pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xgb-fe-codex"
version = "0.1.0"
description = "Experiments comparing XGBoost on raw attributes vs oracle features"
authors = [{name = "AutoGen"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "xgboost>=1.7",
    "optuna>=3.5",
    "PyYAML>=6.0"
]

[project.optional-dependencies]
tests = [
    "pytest>=7.0",
    "pytest-cov>=4.0"
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
